% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbr-nmbayes.R
\name{bbr_nmbayes}
\alias{bbr_nmbayes}
\alias{bbi_nmbayes_model}
\title{NONMEM Bayes models in \code{bbr}}
\description{
This page provides an overview of the basic structure of NONMEM Bayes models
in \code{bbr}. The main entry point for interacting with NONMEM Bayes models is
the \code{bbi_nmbayes_model} object.
}
\details{
\subsection{Model Structure}{

NONMEM models in \code{bbr} have one control stream. Executing the model produces
an output directory. NONMEM Bayes models, on the other hand, have a more
complicated model structure. The user still defines only one control stream.
Underneath, this top-level control stream is then used to derive the control
stream of multiple submodels, one for the initial estimates and one for each
chain.

For a control stream with the base name \code{run},
\link[=nmbayes_submit_model]{submit_model()} will create the following submodels:

\strong{\verb{<run>/init/}} - A model that generates the initial estimates.

\strong{\verb{<run>/<run>-<chain>/}} - A model for each chain that generates samples
using \code{BAYES} or \code{NUTS} estimation methods.
}

\subsection{Defining a NONMEM Bayes model}{

The first step to defining a NONMEM Bayes model will often be to call
\code{\link[=copy_model_as_nmbayes]{copy_model_as_nmbayes()}} on an existing non-Bayes NONMEM model. This creates
the new model with \code{\link[bbr:copy_model_from]{bbr::copy_model_from()}} and then sets the model type to
by "nmbayes". Alternatively you can also create a fresh NONMEM Bayes model
with \code{new_model(..., .model_type = "nmbayes")}. (As with regular NONMEM
models, a control stream must already exist.)

The control stream must define the following two records:
\itemize{
\item a \code{METHOD=CHAIN} estimation record that defines how to generate the
initial estimates for each chain. \code{ISAMPLE} should be set to 0. The
\code{NITER} option controls the number of chains.
\item a \code{METHOD=BAYES} or \code{METHOD=NUTS} estimation record that defines how to
generate the initial estimates for each chain.
}

Here's an example that defines four chains of 500 samples:

\if{html}{\out{<div class="sourceCode">}}\preformatted{$EST METHOD=CHAIN FILE=1100.chn NSAMPLE=4 ISAMPLE=0 SEED=1
     CTYPE=0 IACCEPT=0.3 DF=10 DFS=0

$EST METHOD=NUTS AUTO=2 CTYPE=0 OLKJDF=2 OVARF=1 SEED=1 NBURN=250 NITER=500
     NUTS_DELTA=0.95 PRINT=10 MSFO=./1100.msf RANMETHOD=P PARAFPRINT=10000
     BAYES_PHI_STORE=1
}\if{html}{\out{</div>}}
}
}
