% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmbayes-helpers.R
\name{chain_paths}
\alias{chain_paths}
\title{Return output file from each chain subdirectory}
\usage{
chain_paths(.mod, name = NULL, extension = "")
}
\arguments{
\item{.mod}{A \code{bbi_nmbayes_model} object.}

\item{name}{Name of file, without leading path or extension. If unspecified,
defaults to "\{id\}-\{chain\}".}

\item{extension}{File extension.}
}
\value{
Absolute file paths, one for each chain.
}
\description{
Starting with the top-level NOMEM Bayes model, construct the paths to the
specified per-chain output files.
}
\seealso{
\link{bbr_nmbayes} for a high-level description of how NONMEM Bayes
models are structured
}
