% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-stan-model.R
\name{check_stan_model}
\alias{check_stan_model}
\title{Check Stan model for required files and syntax errors}
\usage{
check_stan_model(.mod, .syntax = TRUE, .error = FALSE)
}
\arguments{
\item{.mod}{A \link{bbi_stan_model} object}

\item{.syntax}{Check the syntax of the model by calling
\link[cmdstanr:model-method-check_syntax]{CmdStanModel$check_syntax()}.}

\item{.error}{If \code{FALSE}, the default, display a message for any problems
found. If \code{TRUE}, signal an error.}
}
\value{
Invisibly return \code{FALSE} if any problems are found (only relevant
when \code{.error} is \code{FALSE}) and \code{TRUE} otherwise.
}
\description{
Ensure that a \code{bbi_stan_model} object has all the files necessary to submit
the model (see \verb{?}\link{bbr_stan} for details). By default, also check the syntax
of that the \code{.stan} file.
}
\details{
Will look for the following:
\itemize{
\item \code{build_path_from_model(.mod, "-standata.R")}
\item \code{build_path_from_model(.mod, "-stanargs.R")}
\item \code{build_path_from_model(.mod, ".stan")}
\item \code{build_path_from_model(.mod, "-init.R")}, unless it is a generated
quantities model (i.e. a \code{bbi_stan_gq_model} subclass).
\item \code{build_path_from_model(.mod, "-fitted-params.R")}, if it is a generated
quantities model (i.e. a \code{bbi_stan_gq_model} subclass).
}
}
