% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-up-to-date.R
\name{check_up_to_date_stan_model}
\alias{check_up_to_date_stan_model}
\alias{check_up_to_date.bbi_stan_model}
\alias{check_up_to_date.bbi_stan_summary}
\title{Check that Stan model is up to date}
\usage{
\method{check_up_to_date}{bbi_stan_model}(.bbi_object, .build_data = TRUE, ...)

\method{check_up_to_date}{bbi_stan_summary}(.bbi_object, .build_data = TRUE, ...)
}
\arguments{
\item{.bbi_object}{the object to check. Could be
a \verb{bbi_{.model_type}_model} object,
a \verb{bbi_{.model_type}_summary} object,
or a \code{bbi_log_df} tibble.}

\item{.build_data}{If \code{TRUE}, the default, run \code{-standata.R} and save the
output to a temp file and check the hash of \emph{the temp file} against the
\code{bbi_config.json} hash. This option actually runs the code and,
importantly, verifies that the input data to \code{-standata.R} has not changed
either. If \code{FALSE}, check the hashes of the \code{-standata.R} and
\code{-standata.json} against the hashes in \code{bbi_config.json} but do \emph{not} run
the \code{-standata.R} script. This option is less secure and primarily exists
for quicker checking if building the data is time consuming for certain
models.}

\item{...}{Arguments passed through (currently none).}
}
\description{
The details of which files are "model files" and which are "data files" are
detailed in the \code{\link[bbr:check_up_to_date]{bbr::check_up_to_date()}} docs.
}
\details{
\itemize{
\item The model files:
\itemize{
\item \verb{<run>.stan}
\item \verb{<run>-init.R}
\item \verb{<run>-stanargs.R}
}
\item The data files (see \code{.build_data} argument):
\itemize{
\item \verb{<run>-standata.R}
\item \verb{<run>-standata.json}
}
}
}
