% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy-model-from.R
\name{copy_model_as_nmbayes}
\alias{copy_model_as_nmbayes}
\title{Copy a "regular" NONMEM model to NONMEM Bayes model}
\usage{
copy_model_as_nmbayes(
  .parent_mod,
  .new_model = NULL,
  .description = NULL,
  .based_on_additional = NULL,
  .add_tags = NULL,
  .star = NULL,
  .inherit_tags = FALSE,
  .update_model_file = TRUE,
  .overwrite = FALSE
)
}
\arguments{
\item{.parent_mod}{A \code{bbi_nonmem_model} object to copy. This should \emph{not} be
a \code{bbi_nmbayes_model} subclass; in that case, use \code{\link[bbr:copy_model_from]{bbr::copy_model_from()}}
to copy the model in the standard way.}

\item{.new_model}{Path to the new model, either absolute or relative to the
path to \code{.parent_mod}. Represents an absolute model path, which is the path
to the YAML file and model file, both without extension, and the output
directory (once the model is run). Numeric values will be coerced to
character. If \code{NULL}, the default, will try to increment to the next integer
in the destination directory. See examples for usage.}

\item{.description}{Character scalar description of new model run. This will
be stored in the yaml (and can be viewed later in \code{run_log()}).}

\item{.based_on_additional}{Character vector of path(s) to other models that
this model was "based on." These are used to reconstuct model developement
and ancestry. \strong{Paths must be relative to \code{.new_model} path.} Note that
the \code{.parent_model} will automatically be added to the \code{based_on} field, so
no need to include that here.}

\item{.add_tags}{Character vector with any new tags(s) to be added to
\verb{{.new_model}.yaml}}

\item{.star}{Boolean, marks model to indicate special interest level.}

\item{.inherit_tags}{If \code{FALSE}, the default, new model will only have any
tags passed to \code{.add_tags} argument. If \code{TRUE} inherit any tags from
\code{.parent_mod}, with any tags passed to \code{.add_tags} appended.}

\item{.update_model_file}{Whether to update the newly created model file. If
\code{TRUE}, the model will be adjusted by \code{\link[bbr:copy_model_from]{bbr::copy_model_from()}}. In
addition, any existing estimation records will be deleted, and the two
estimation records required for nmbayes models will be added along with a
comment.}

\item{.overwrite}{If \code{FALSE}, the default,  function will error if a model
file already exists at specified \code{.new_model} path. If \code{TRUE} any existing
file at \code{.new_model} will be overwritten silently.}
}
\value{
A \code{bbi_nmbayes_model} object for the new model.
}
\description{
Like \code{\link[bbr:copy_model_from]{bbr::copy_model_from()}}, but switch the model type from "nonmem" to
"nmbayes" to begin defining a \link[=bbr_nmbayes]{NONMEM Bayes model}.
}
\seealso{
\link{bbr_nmbayes} for a high-level description of how NONMEM Bayes
models are structured in bbr
}
