% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy-model-from.R
\name{copy_stan_model_as_gq}
\alias{copy_stan_model_as_gq}
\title{Copy a "regular" Stan model to a generated quantities model}
\usage{
copy_stan_model_as_gq(
  .parent_mod,
  .new_model = NULL,
  .description = NULL,
  .based_on_additional = NULL,
  .add_tags = NULL,
  .star = NULL,
  .inherit_tags = FALSE,
  .overwrite = FALSE
)
}
\arguments{
\item{.parent_mod}{A \code{bbi_stan_model} object to copy. This should \emph{not} be a
\code{bbi_stan_gq_model} subclass; in that case, use \code{\link[bbr:copy_model_from]{bbr::copy_model_from()}} to
copy the model in the standard way.}

\item{.new_model}{Path to the new model, either absolute or relative to the
path to \code{.parent_mod}. Represents an absolute model path, which is the path
to the YAML file and model file, both without extension, and the output
directory (once the model is run). Numeric values will be coerced to
character. If \code{NULL}, the default, will try to increment to the next integer
in the destination directory. See examples for usage.}

\item{.description}{Character scalar description of new model run. This will
be stored in the yaml (and can be viewed later in \code{run_log()}).}

\item{.based_on_additional}{Character vector of path(s) to other models that
this model was "based on." These are used to reconstuct model developement
and ancestry. \strong{Paths must be relative to \code{.new_model} path.} Note that
the \code{.parent_model} will automatically be added to the \code{based_on} field, so
no need to include that here.}

\item{.add_tags}{Character vector with any new tags(s) to be added to
\verb{{.new_model}.yaml}}

\item{.star}{Boolean, marks model to indicate special interest level.}

\item{.inherit_tags}{If \code{FALSE}, the default, new model will only have any
tags passed to \code{.add_tags} argument. If \code{TRUE} inherit any tags from
\code{.parent_mod}, with any tags passed to \code{.add_tags} appended.}

\item{.overwrite}{If \code{FALSE}, the default,  function will error if a model
file already exists at specified \code{.new_model} path. If \code{TRUE} any existing
file at \code{.new_model} will be overwritten silently.}
}
\description{
Like \code{\link[bbr:copy_model_from]{bbr::copy_model_from()}}, but switch the model type from "stan" to
"stan_gq" begin defining a model for standalone generated quantities.
}
\details{
Aside from adjusting the \code{model_type} value in the new model's YAML file,
this function does the following:
\itemize{
\item adds a \verb{<run>-fitted-param.R} file. Its value \emph{may} be suitable for your
purposes, but please inspect it and edit as needed.
\item copies \verb{<run>.stan} and \verb{<run>-standata.R} as is from the parent model. At
least \verb{<run>.stan} will need to be manually modified before running the
model.
\item \verb{<run>-init.R} is \emph{not} copied because it is not relevant for standalone
generated quantities.
\item Only the \code{seed} argument from the parent model's \verb{<run>-stanargs.R} is
copied to the new model's \verb{<run>-stanargs.R}.
}
}
\seealso{
\link{bbr_stan} for a high-level description of how Stan models are
structured
}
