% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-gq-parent.R
\name{find_stan_gq_children}
\alias{find_stan_gq_children}
\title{Search for models that have a given model as the gq_parent}
\usage{
find_stan_gq_children(.mod, .base_dir = NULL, .recurse = FALSE)
}
\arguments{
\item{.mod}{A \code{bbi_stan_model} object.}

\item{.base_dir}{Base directory to look in for models. If \code{NULL} (the
default), look for models in the directory that contains \code{.mod}.}

\item{.recurse}{If \code{FALSE}, the default, only include models in \code{.base_dir}
but \emph{not} subdirectories. This is passed through to \code{fs::dir_ls()} --
If \code{TRUE} recurse fully, if a positive number the number of levels to recurse.}
}
\description{
By looking at the "gq_parent" field of a "stan_gq" model, you can identify
which "stan" models are used as input. This function helps you go in the
other direction: find which "stan_gq" models under a directory point to the
specified "stan" model.
}
\seealso{
\code{\link[=get_stan_gq_parent]{get_stan_gq_parent()}} and \link{modify_stan_gq_parent} for getting and
modifying the "gq_parent" value of a \code{bbi_stan_gq_model} object, \link{bbr_stan}
for more information about standalone generated quantities
}
