% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanargs.R
\name{get_stanargs}
\alias{get_stanargs}
\title{Return the current stanargs attached to a model}
\usage{
get_stanargs(.mod)
}
\arguments{
\item{.mod}{The model object to submit.}
}
\description{
Return the current stanargs attached to a model
}
