% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-diff.R
\name{model_diff.bbi_stan_model}
\alias{model_diff.bbi_stan_model}
\title{Compare Stan models}
\usage{
\method{model_diff}{bbi_stan_model}(
  .mod,
  .mod2 = NULL,
  .file = c("model", "standata", "init", "stanargs"),
  ...,
  .viewer = FALSE
)
}
\arguments{
\item{.mod, .mod2}{Show how \code{.mod2} has changed compared \code{.mod}. If \code{mod2} is
\code{NULL}, compare \code{.mod} to the model returned by \code{get_based_on(.mod)}.}

\item{.file}{Which files to compare:
\itemize{
\item model: \verb{<run>.stan}
\item standata: \verb{<run>-standata.R}
\item init: \verb{<run>-init.R}
\item stanargs: \verb{<run>-stanargs.R}
}}

\item{...}{arguments passed through to methods. (Currently none.)}

\item{.viewer}{If \code{FALSE}, the default, prints diff to console or renders in
Rmd. If \code{TRUE}, render the diff in the Viewer window. Note: this option
will hold the console with \verb{Press ENTER to continue...} so it is only
recommended for interactive use.}
}
\description{
Diff files between \code{bbi_stan_model} objects.
}
\seealso{
\code{\link[bbr:model_diff]{bbr::model_diff()}}
}
