% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-gq-parent.R
\name{modify_stan_gq_parent}
\alias{modify_stan_gq_parent}
\alias{add_stan_gq_parent}
\alias{replace_all_stan_gq_parent}
\alias{remove_stan_gq_parent}
\title{Modify gq_parent of a standalone generated quantities model}
\usage{
add_stan_gq_parent(.mod, .parent)

replace_all_stan_gq_parent(.mod, .parent)

remove_stan_gq_parent(.mod, .parent)
}
\arguments{
\item{.mod}{A \code{bbi_stan_gq_model} object.}

\item{.parent}{Character vector of paths to parent model. If relative, the
paths are taken as relative to the directory containing \code{.mod}.}
}
\value{
The modified model objects (after writing changes to disk).
}
\description{
Modify gq_parent of a standalone generated quantities model
}
