% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-join-bayes.R
\name{nm_join_bayes}
\alias{nm_join_bayes}
\alias{nm_join_bayes_quick}
\title{Join input data to Bayesian model output summaries}
\usage{
nm_join_bayes_quick(
  mod,
  data = NULL,
  join_col = "NUM",
  point_fn = stats::median
)

nm_join_bayes(
  mod,
  mod_mrgsolve,
  data = NULL,
  join_col = "NUM",
  dv_col = "DV",
  y_col = "Y",
  point_fn = stats::median,
  probs = c(0.025, 0.975),
  resid_var = TRUE,
  n_post = 1000,
  log_dv = FALSE,
  epred = TRUE,
  ipred = TRUE,
  ipred_path = NULL,
  ewres_npde = FALSE,
  npde_decorr_method = c("cholesky", "inverse", "polar"),
  min_batch_size = 200
)
}
\arguments{
\item{mod}{A \code{bbi_nmbayes_model} object.}

\item{data}{A NONMEM dataset for \code{mod} to use instead of the default one
extracted from \code{mod} with \code{\link[bbr:nm_file]{bbr::nm_data()}}. This data frame must include a
\code{join_col} column.}

\item{join_col}{Column used to join \code{data} to a data frame constructed by
combining values from the \code{bbr-bayes-join.tab} file in each chain
subdirectory.}

\item{point_fn}{Function used to calculate point estimates of table values
across chains and of simulated EPRED and IPRED values (e.g., mean or
median).}

\item{mod_mrgsolve}{An mrgsolve model object, potentially updated to be
optimized for simulation from the data set and model (e.g., ODE solver
tolerance). This must capture \code{y_col}.}

\item{dv_col}{Pass this \code{data} column as the dependent variable when
calculating EWRES and NPDE.}

\item{y_col}{The name of the dependent variable in \code{mod_mrgsolve}. This is a
simulated quantity corresponding to \code{dv_col}.}

\item{probs}{A two-item vector of lower and upper probabilities to pass to
\code{\link[stats:quantile]{stats::quantile()}} to calculate the bounds of the simulated EPRED and
IPRED values.}

\item{resid_var}{Whether to include residual variability in simulations.}

\item{n_post}{Randomly select this number of posterior draws to use as input
to the simulation.}

\item{log_dv}{Whether the DV was log-transformed.}

\item{epred}{Simulate EPRED values, including the point estimate as \code{EPRED}
and the bounds as \code{EPRED_lo} and \code{EPRED_hi}.}

\item{ipred}{Simulate IPRED values, including the point estimate as \code{IPRED}
and the bounds as \code{IPRED_lo} and \code{IPRED_hi}.}

\item{ipred_path}{Write the IPRED simulation result (as comma-separated
values) to this path. This is useful if you need access to the full results
(e.g., for LOO calculations), not just the summary returned by this
function.}

\item{ewres_npde}{Whether to replace EWRES and NPDE values obtained from
table with ones generated with \pkg{npde}.}

\item{npde_decorr_method}{Pass this value to \code{decorr.method} of
\code{\link[npde:autonpde]{npde::autonpde()}}.}

\item{min_batch_size}{To simulate EPRED and IPRED values, posterior samples
are split into batches and sent as items to \pkg{future} \verb{*apply}
functions. The number of batches is chosen so that each batch has at least
the number of samples specified by this argument.

\strong{Note:} You may be able to tune this value to get better performance for
a particular case (considering factors such as the model, targeted future
backend, and machine). However, the results across runs with the same
initial seed will differ if you change this value. To avoid the results
varying across different machines and future backends, do not dynamically
set this value based on the number of available workers.}
}
\value{
A data frame. The base data frame is the result combining the
\code{bbr-bayes-join.tab} values for each chain, collapsing across \code{join_col}
with \code{point_fn}, and then joining \code{data}. \code{EPRED}, \code{IPRED}, \code{EWRES}, and
\code{NPDE} values are replaced with a simulated estimate, if requested by the
corresponding argument.
}
\description{
\code{nm_join_bayes()} and \code{nm_join_bayes_quick()} both join input data to model
output summaries. Conceptually these are similar to \code{\link[bbr:nm_join]{bbr::nm_join()}}, but
they summarize across chains.

\code{nm_join_bayes()} selects a subset of posterior samples and \strong{simulates}
EPRED and IPRED with specified mrgsolve model. These simulated values are
passed to the specified summary function (\code{\link[stats:median]{stats::median()}} by default).

\code{nm_join_bayes_quick()}, on the other hand, avoids the simulation and passes
the the table values for each chain to the summary function. \strong{Warning}:
this estimate should not be considered as reliable but may be useful in the
early stages of model development.
}
\details{
HELP: (details)
}
\seealso{
\code{\link[bbr:nm_join]{bbr::nm_join()}}
}
