% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-nonmem-finished.R
\name{nmbayes_check_nonmem_finished}
\alias{nmbayes_check_nonmem_finished}
\alias{check_nonmem_finished.bbi_nmbayes_model}
\title{Check if a NONMEM Bayes model has finished running}
\usage{
\method{check_nonmem_finished}{bbi_nmbayes_model}(.mod, ...)
}
\arguments{
\item{.mod}{A \code{bbi_nmbayes_model} object.}

\item{...}{Arguments passed to method call on each chain submodel.}
}
\value{
\code{TRUE} if all of the chain models have finished, and \code{FALSE}
otherwise.
}
\description{
This function implements the \code{\link[bbr:check_nonmem_finished]{bbr::check_nonmem_finished()}} method for NONMEM
Bayes models. To decide if a model is finished, it runs
\code{\link[bbr:check_nonmem_finished]{bbr::check_nonmem_finished()}} on each of the chain submodels.
}
\details{
As with regular NONMEM models, the model is \emph{not} considered finished if the
output directory does not exist.
}
