% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmbayes-convert.R
\name{nmbayes_convert}
\alias{nmbayes_convert}
\title{Convert a non-Bayes NONMEM model into a nmbayes model}
\usage{
nmbayes_convert(.mod)
}
\arguments{
\item{.mod}{A \code{bbi_nonmem_model} object.}
}
\value{
The converted model. If \code{.mod} is already a \code{bbi_nmbayes_model}, it
is returned as is.
}
\description{
NONMEM Bayes models have a different type than other Bayes models because
they have a more complicated structure underneath that requires special
handling. This method converts the type of an existing model from "nonmem" to
"nmbayes" by modifying the underlying YAML file. A call to
\code{\link[bbr:copy_model_from]{bbr::copy_model_from()}} should usually precede a call to this function.
}
