% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit-model.R
\name{nmbayes_submit_model}
\alias{nmbayes_submit_model}
\alias{submit_model.bbi_nmbayes_model}
\title{Submit model based on a \code{bbi_nmbayes_model} object}
\usage{
\method{submit_model}{bbi_nmbayes_model}(
  .mod,
  .bbi_args = NULL,
  .mode = getOption("bbr.bbi_exe_mode"),
  ...,
  .overwrite = NULL,
  .config_path = NULL,
  .wait = TRUE,
  .dry_run = FALSE
)
}
\arguments{
\item{.mod}{The model object to submit.}

\item{.bbi_args}{A named list specifying arguments to pass to bbi
formatted like \code{list("nm_version" = "nm74gf_nmfe", "json" = T, "threads" = 4)}. Run \code{\link[bbr:print_bbi_args]{print_bbi_args()}} to see valid arguments. Note that bbr does
not support changing the output directory (including through the model or
global YAML files).}

\item{.mode}{Either \code{"sge"}, the default, to submit model(s) to the grid or
\code{"local"} for local execution. This can be passed directly to this argument
or set globally with \code{options("bbr.bbi_exe_mode")}.}

\item{...}{args passed through to \code{bbi_exec()}}

\item{.overwrite}{Logical to specify whether or not to overwrite existing
model output from a previous run. If \code{NULL}, the default, will defer to
setting in \code{.bbi_args} or \code{bbi.yaml}. If \emph{not} \code{NULL} will override any
settings in \code{.bbi_args} or \code{bbi.yaml}.}

\item{.config_path}{Path to a bbi configuration file. If \code{NULL}, the
default, will attempt to use a \code{bbi.yaml} in the same directory as the
model.}

\item{.wait}{If \code{TRUE}, the default, wait for the bbi process to return
before this function call returns. If \code{FALSE} function will return while
bbi process runs in the background.}

\item{.dry_run}{Do not submit the sampling runs; just report what command
would be executed via the returned object. \strong{Note}: The METHOD=CHAIN model
is executed to generate the initialization values regardless of this value.}
}
\description{
Model submission consists of two steps: generating the initial values
(\code{METHOD=CHAIN} run) and sampling for each chain (\code{METHOD=BAYES} or
\code{METHOD=NUTS} runs).
}
\seealso{
\link{bbr_nmbayes} for a high-level description of how NONMEM Bayes
models are structured
}
