% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmbayes-tail.R
\name{nmbayes_tail}
\alias{nmbayes_tail}
\alias{nmbayes_tail_lst}
\alias{nmbayes_tail_output}
\alias{tail_lst.bbi_nmbayes_model}
\alias{tail_output.bbi_nmbayes_model}
\title{Inspect first and last lines of select files from a NONMEM Bayes model}
\usage{
\method{tail_lst}{bbi_nmbayes_model}(.mod, .head = 3, .tail = 5, .print = TRUE, .return = FALSE, ...)

\method{tail_output}{bbi_nmbayes_model}(.mod, .head = 3, .tail = 5, .print = TRUE, .return = FALSE, ...)
}
\arguments{
\item{.mod}{Model object or path associated with file.}

\item{.head}{Integer for number of lines to read from the top of the file}

\item{.tail}{Integer for number of lines to read from the bottom of the file}

\item{.print}{If \code{TRUE}, the default, print resulting head and tail to console.}

\item{.return}{If \code{TRUE}, return the head and tail results for each chain.
See \sQuote{Value} section for more details.}

\item{...}{Arguments passed to method call on each chain submodel.}
}
\value{
If \code{.return} is \code{TRUE}, return a list of character vectors, with an
element for each chain. The character vector is comprised of the first and
last lines of the file under a chain subdirectory. For \code{tail_output()}, the
element will be \code{NULL} if 1) the model has started to run but hasn't yet
created the \code{OUTPUT} file or 2) the model has finished running.
}
\description{
\code{tail_lst()} and \code{tail_output()} provide \code{\link[bbr:check_file]{bbr::tail_lst()}} and
\code{\link[bbr:check_file]{bbr::tail_output()}} methods for \code{bbi_nmbayes_model} objects. Compared to the
methods for \code{bbi_nonmem_model} objects, the key difference is that these
operate on the \verb{*.lst} or \code{OUTPUT} file of each chain submodel.
}
\seealso{
\code{\link[=chain_paths]{chain_paths()}} for constructing chain submodels paths
}
