% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{open_model_file}
\alias{open_model_file}
\alias{open_stan_file}
\alias{open_standata_file}
\alias{open_init_file}
\title{Open Stan model file in a text editor}
\usage{
open_stan_file(.mod)

open_standata_file(.mod)

open_init_file(.mod)
}
\arguments{
\item{.mod}{a \code{bbi_stan_model}}
}
\description{
These light wrappers around \code{\link[bbr:build_path_from_model]{bbr::build_path_from_model()}} enable quickly
jumping to a model file in RStudio or any other editor supported by
\code{\link[utils:file.edit]{utils::file.edit()}}.
}
\section{Functions}{
\itemize{
\item \code{open_stan_file()}: Open \verb{<run>.stan}.

\item \code{open_standata_file()}: Open \verb{<run>-standata.R}.

\item \code{open_init_file()}: Open \verb{<run>-init.R}.

}}
