% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{open_stan_file}
\alias{open_stan_file}
\alias{open_stanmod_file}
\alias{open_standata_file}
\alias{open_staninit_file}
\alias{open_stan_fitted_params_file}
\title{Open Stan model file in a text editor}
\usage{
open_stanmod_file(.mod)

open_standata_file(.mod)

open_staninit_file(.mod)

open_stan_fitted_params_file(.mod)
}
\arguments{
\item{.mod}{a \code{bbi_stan_model}}
}
\description{
These light wrappers around \code{\link[bbr:build_path_from_model]{bbr::build_path_from_model()}} enable quickly
jumping to a model file in RStudio or any other editor supported by
\code{\link[utils:file.edit]{utils::file.edit()}}.
}
\section{Functions}{
\itemize{
\item \code{open_stanmod_file()}: Open \verb{<run>.stan}.

\item \code{open_standata_file()}: Open \verb{<run>-standata.R}.

\item \code{open_staninit_file()}: Open \verb{<run>-init.R} (not relevant for standalone
generated quantities).

\item \code{open_stan_fitted_params_file()}: Open \verb{<run>-fitted-params.R} (only relevant for
standalone generated quantities).

}}
