% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-sims.R
\name{run_sims}
\alias{run_sims}
\title{TODO}
\usage{
run_sims(
  mod,
  mod_mrgsolve,
  data = NULL,
  ext = NULL,
  join_col = "NUM",
  y_col = "Y",
  point_fn = stats::median,
  probs = c(0.025, 0.975),
  resid_var = TRUE,
  n_post = 1000,
  log_dv = FALSE,
  epred = TRUE,
  ipred = TRUE,
  ipred_path = NULL,
  ewres_npde = FALSE
)
}
\arguments{
\item{mod}{A bbi_nmbayes_model object.}

\item{mod_mrgsolve}{TODO}

\item{data}{TODO}

\item{ext}{TODO}

\item{join_col}{TODO}

\item{y_col}{TODO}

\item{resid_var}{TODO}

\item{n_post}{TODO}

\item{log_dv}{TODO}

\item{epred}{TODO}

\item{ipred}{TODO}

\item{ipred_path}{TODO}

\item{ewres_npde}{TODO}

\item{point_est}{TODO}

\item{ci_level}{TODO}
}
\value{
TODO
}
\description{
TODO
}
