% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanargs.R
\name{set_stanargs}
\alias{set_stanargs}
\title{Set arguments to be passed to cmdstanr$sample()}
\usage{
set_stanargs(.mod, .stanargs, .clear = FALSE)
}
\arguments{
\item{.mod}{A \code{bbi_stan_model} object.}

\item{.stanargs}{args passed through to \code{\link[cmdstanr:model-method-sample]{cmdstanr::sample()}}.}

\item{.clear}{If \code{FALSE}, the default, add any new args and only overwrite
existing args that are passed in through \code{.stanargs}. If \code{TRUE}, overwrite
\emph{all} attached args so that only \code{.stanargs} will be attached.}
}
\description{
Set arguments to be passed to cmdstanr$sample()
}
