% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkage.R
\name{shrinkage}
\alias{shrinkage}
\alias{shrinkage.bbi_nmbayes_model}
\alias{shrinkage.draws}
\alias{shrinkage.rvar}
\title{Calculate shrinkage in variance of errors}
\usage{
shrinkage(errors, ...)

\method{shrinkage}{bbi_nmbayes_model}(errors, ...)

\method{shrinkage}{draws}(
  errors,
  errors_name,
  variance_name = NULL,
  from_diag = FALSE,
  group_idx = NULL,
  ...
)

\method{shrinkage}{rvar}(errors, variance = NULL, group_idx = NULL, ...)
}
\arguments{
\item{errors}{An object from which errors can be extracted.}

\item{...}{Additional arguments passed on to methods.}

\item{errors_name}{Name of a parameter to extract from the \code{draws_rvars}
object and use as the errors.}

\item{variance_name}{Name of a parameter to extract from the \code{draws_rvars}
object and use as the variance. The dimensions of this value should
correspond to the dimenions in the errors after summarizing across the
group dimension(s). If no name is provided, the variance across groups is
calculated for the extracted errors.}

\item{from_diag}{To obtain the variance values, take the diagonal of the
matrix return by \code{variance_name}.}

\item{group_idx}{A vector of indices specifying which dimension(s) correspond
to groups. Defaults to the last dimension when not specified.}

\item{variance}{A \pkg{posterior} rvar object specifying variance of the
errors. If not specified, it is set to the variance across groups in
\code{errors}.}
}
\value{
Vector of shrinkage estimates, in the same order as the values
specified by \code{errors}.

The dimensions of the input errors determines the dimensions of the return
value (dropping any dimensions with only one level). For example, errors of
the form \code{X[i,j]}, where \code{j} is the group-level index, would lead to a
vector of \code{i} values. And errors \code{X[i,j,k]}, where \code{k} is the group-level
index, would lead to a matrix with \code{i} rows and \code{j} columns.
}
\description{
Using the posterior samples, estimate how much the variance of errors is
reduced by group-level pooling.

For a \code{bbi_nmbayes_model} object, the \code{ETA} values are taken as the errors,
and variance for each \code{ETA} is extracted from the diagonal of the \code{OMEGA}
matrix. As a special case, if \verb{*.iph} files do not exist, the shrinkage
values are collected from the \verb{*.shk} files.

For a \pkg{posterior} draws object, the errors and variance are extracted for
the specified parameter names.

For a \pkg{posterior} rvar object, the errors and variance are supplied
directly.
}
\details{
The shrinkage calculation here follows Gelman and Pardoe's pooling factor
definition, with a few adjustments:
\itemize{
\item Keeping with pharmacometrics conventions, the calculation uses standard
deviation rather than variance.
\item As with NONMEM's shrinkage values, the returned values are percentages
rather than fractions.
\item If the variance of the errors is supplied, the denominator is calculated
by taking the expectation of those values.
}
}
\references{
Andrew Gelman, Iain Pardoe (2006) Bayesian measures of explained
variance and pooling in multilevel (hierarchical) models. \emph{Technometrics}.
48(2), 241--251.
}
