% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit-model.R
\name{stan_submit_model}
\alias{stan_submit_model}
\alias{submit_model.bbi_stan_model}
\title{Submit model based on a \code{bbi_stan_model} object}
\usage{
\method{submit_model}{bbi_stan_model}(.mod, .mode = c("local"), ..., .overwrite = NULL)
}
\arguments{
\item{.mod}{The model object to submit.}

\item{.mode}{Mode of model submission. Stan models currently only support
local execution.}

\item{...}{Additional arguments (ignored for all Stan models).}

\item{.overwrite}{Logical to specify whether or not to overwrite existing
model output from a previous run. If \code{NULL}, the default, will defer to
setting in \code{.bbi_args} or \code{bbi.yaml}. If \emph{not} \code{NULL} will override any
settings in \code{.bbi_args} or \code{bbi.yaml}.}
}
\description{
The model is executed via \code{\link[cmdstanr:model-method-sample]{cmdstanr::sample()}}.
}
