% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-summary-log.R
\name{stan_summary_log}
\alias{stan_summary_log}
\alias{stan_add_summary}
\title{Create a tibble summarizing results for a set of Stan models}
\usage{
stan_summary_log(
  .base_dir,
  .recurse = FALSE,
  .include = NULL,
  variables = "lp__",
  summary_fns = NULL
)

stan_add_summary(.log_df, variables = "lp__", summary_fns = NULL)
}
\arguments{
\item{.base_dir}{Base directory to look in for models.}

\item{.recurse}{If \code{FALSE}, the default, only include models in \code{.base_dir}
but \emph{not} subdirectories. This is passed through to \code{fs::dir_ls()} --
If \code{TRUE} recurse fully, if a positive number the number of levels to recurse.}

\item{.include}{A character vector specifying which runs or model tags to include in the run log.}

\item{variables}{Model variables to summarize with \code{summary_fns}. This value
is passed through to the \code{CmdStanFit}
\link[cmdstanr:fit-method-summary]{$summary()} method. To summarize these
variables, their draws will be read in. It is valid to specify a variable
that only exists in a subset of the models.

By default this is restricted to "lp__", but you will likely want to extend
this with a few variables of interest. \verb{$summary()} will use \emph{all}
variables if you pass \code{NULL}, but it's recommended to keep the set small
for performance reasons and because each variable adds a number of columns
equal to the number of \code{summary_fns} (or possibly more, if a summary
function returns a named vector). To disable extracting and summarizing the
draws entirely, set \code{summary_fns} to an empty list.}

\item{summary_fns}{A list of summary or diagnostic functions passed to
\link[=draws_summary]{posterior::summarize_draws()}. The result will be included
as a new column, \verb{\{variable\}_\{name\}}, where "name" is determined by the
names of the returned vector or otherwise by the name of the \code{summary_fns}
item.

With the default value of \code{NULL}, the following functions are used:
\code{\link[=mean]{mean()}}, \code{\link[stats:median]{stats::median()}}, \code{\link[posterior:quantile2]{posterior::quantile2()}}, \code{\link[posterior:rhat]{posterior::rhat()}},
\code{\link[posterior:ess_bulk]{posterior::ess_bulk()}}, and \code{\link[posterior:ess_tail]{posterior::ess_tail()}}. Pass an empty list to
disable all variable summaries.}

\item{.log_df}{a \code{bbi_run_log_df} tibble (the output of \code{\link[bbr:run_log]{run_log()}})}
}
\value{
A tibble. For \code{stan_summary_log()}, the tibble will have a row for
each Stan model collected from under the directory. For
\code{stan_add_summary()}, the tibble will have the same number of rows as
\code{.log_df} with additional columns.
}
\description{
There are two entry points for generating the summary log:
\itemize{
\item \code{stan_summary_log()}: return a tibble the summarizes the Stan models under
the specified directory
\item \code{stan_add_summary()}: join summary columns to a \code{\link[bbr:run_log]{bbr::run_log()}} tibble.
At least one row of the run log tibble must have a Stan model.
}
}
\details{
The information for each row is primarily extracted from a few methods of the
\pkg{cmdstanr} fit object: \link[cmdstanr:fit-method-metadata]{$metadata},
\link[cmdstanr:fit-method-diagnostic_summary]{$diagnostic_summary()}, and
\link[cmdstanr:fit-method-summary]{$summary()}.
}
\seealso{
\url{https://mc-stan.org/misc/warnings} for an overview of diagnostics
and warnings
}
