% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-file-to-model-dir.R
\name{stanargs_scaffold}
\alias{stanargs_scaffold}
\title{Creates a starter -stanargs.R with only seed in it}
\usage{
stanargs_scaffold(.mod)
}
\description{
Creates a starter -stanargs.R with only seed in it
}
