% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-stan-model.R
\name{check_stan_model}
\alias{check_stan_model}
\title{Checks a \code{bbi_stan_model} for necessary files}
\usage{
check_stan_model(.mod, .error = FALSE)
}
\arguments{
\item{.mod}{A \link{bbi_stan_model} object}

\item{.error}{If \code{FALSE}, the default, will warn if any necessary files are
missing. If \code{TRUE} will error instead.}
}
\value{
Invisibly return \code{FALSE} if any problems are found (only relevant
when \code{.error} is \code{FALSE}) and \code{TRUE} otherwise.
}
\description{
Checks a \code{bbi_stan_model} object to make sure it can find all of the files
necessary to submit the model. (See \verb{?}\link{bbr_stan} for details about these
files.)
}
\details{
Will look for the following:
\itemize{
\item \code{build_path_from_model(.mod, "-standata.R")}
\item \code{build_path_from_model(.mod, "-stanargs.R")}
\item \code{build_path_from_model(.mod, ".stan")}
\item \code{build_path_from_model(.mod, "-init.R")}, unless it is a generated
quantities model (i.e. a \code{bbi_stan_gq_model} subclass).
\item \code{build_path_from_model(.mod, "-fitted-params.R")}, if it is a generated
quantities model (i.e. a \code{bbi_stan_gq_model} subclass).
}
}
