% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-gq-parent.R
\name{modify_stan_gq_parent}
\alias{modify_stan_gq_parent}
\alias{add_stan_gq_parent}
\alias{replace_all_stan_gq_parent}
\alias{remove_stan_gq_parent}
\title{Modify gq_parent of a standalone generated quantities model}
\usage{
add_stan_gq_parent(.mod, .parent)

replace_all_stan_gq_parent(.mod, .parent)

remove_stan_gq_parent(.mod, .parent)
}
\arguments{
\item{.mod}{A \code{bbi_stan_gq_model} object.}

\item{.parent}{Character vector of paths to parent model. If relative, the
paths are taken as relative to the directory containing \code{.mod}.}
}
\value{
The modified model objects (after writing changes to disk).
}
\description{
In most cases, the "gq_parent" field is automatically handled by
\code{\link[=copy_model_as_stan_gq]{copy_model_as_stan_gq()}}. However, the following convenience functions
enable manually modifying the value if needed:
\itemize{
\item \code{add_stan_gq_parent()}: add specified values, keeping any existing values
\item \code{replace_all_stan_gq_parent()}: replace all existing values with the
specified values
\item \code{remove_stan_gq_parent()}: remove the specified values
}
}
\seealso{
\link{get_stan_gq_parent} for getting the "gq_parent" value, \link{bbr_stan}
for more information about standalone generated quantities
}
