% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-join-bayes.R
\name{nm_join_bayes}
\alias{nm_join_bayes}
\title{Join Bayesian model output summaries to input data}
\usage{
nm_join_bayes(
  .mod,
  mod_mrgsolve,
  .join_col = "NUM",
  .files = NULL,
  .superset = FALSE,
  ...,
  dv_col = "DV",
  y_col = "Y",
  point_fn = stats::median,
  probs = c(0.025, 0.975),
  resid_var = TRUE,
  n_post = 1000,
  epred = TRUE,
  ipred = TRUE,
  ipred_path = NULL,
  ewres_npde = TRUE,
  npde_decorr_method = c("cholesky", "inverse", "polar"),
  presim_fn = NULL,
  min_batch_size = 200
)
}
\arguments{
\item{.mod}{A \code{bbi_nmbayes_model} object.}

\item{mod_mrgsolve}{An mrgsolve model object, potentially updated to be
optimized for simulation from the data set and model (e.g., ODE solver
tolerance). This must capture \code{y_col}.}

\item{.join_col}{Use this column to join tables files to the input data.}

\item{.files}{Paths to table files to pass to \code{\link[bbr:nm_join]{bbr::nm_join()}} calls (on per
each chain submodel). By default, all tables specified in the \verb{$TABLE}
blocks of the control stream will be used. Note that, unlike
\code{\link[bbr:nm_join]{bbr::nm_join()}}, this function accepts only relative paths because the
paths need to be valid for each chain submodel.}

\item{.superset}{If \code{FALSE}, the default, the data will be joined to the
NONMEM output and if \code{TRUE}, the NONMEM output will be joined to the data;
that is, if you use \code{.superset}, you will get the same number of rows as
you have in the input data and NONMEM output columns like \code{PRED} and
\code{CWRES} will be filled with \code{NA}.}

\item{...}{Additional arguments passed to \code{\link[bbr:nm_join]{bbr::nm_join()}}.}

\item{dv_col}{Pass this data column as the dependent variable when
calculating EWRES and NPDE.}

\item{y_col}{The name of the dependent variable in \code{mod_mrgsolve}. This is a
simulated quantity corresponding to \code{dv_col}.}

\item{point_fn}{Function used to calculate point estimates of table values
across chains and of simulated EPRED and IPRED values (e.g., mean or
median).}

\item{probs}{A two-item vector of lower and upper probabilities to pass to
\code{\link[stats:quantile]{stats::quantile()}} to calculate the bounds of the simulated EPRED and
IPRED values.}

\item{resid_var}{Whether to include residual variability in simulations.}

\item{n_post}{Randomly select this number of posterior draws to use as input
to the simulation.}

\item{epred}{Simulate EPRED values, including the point estimate as \code{EPRED}
and the bounds as \code{EPRED_lo} and \code{EPRED_hi}.}

\item{ipred}{Simulate IPRED values, including the point estimate as \code{IPRED}
and the bounds as \code{IPRED_lo} and \code{IPRED_hi}.}

\item{ipred_path}{Write the IPRED simulation result (as comma-separated
values) to this path. This is useful if you need access to the full results
(e.g., for LOO calculations), not just the summary returned by this
function.}

\item{ewres_npde}{Whether to replace EWRES and NPDE values obtained from
table with ones generated with \pkg{npde}.}

\item{npde_decorr_method}{Pass this value to \code{decorr.method} of
\code{\link[npde:autonpde]{npde::autonpde()}}.}

\item{presim_fn}{Before simulating, apply this function to the data frame
that results from joining the input data and table values. The main purpose
of this argument is to provide a way to do any column renames that are
required for the mrgsolve simulation (e.g., renaming a column to "TIME").}

\item{min_batch_size}{To simulate EPRED and IPRED values, posterior samples
are split into batches and sent as items to \pkg{future} \verb{*apply}
functions. The number of batches is chosen so that each batch has at least
the number of samples specified by this argument.

\strong{Note:} You may be able to tune this value to get better performance for
a particular case (considering factors such as the model, targeted future
backend, and machine). However, the results across runs with the same
initial seed will differ if you change this value. To avoid the results
varying across different machines and future backends, do not dynamically
set this value based on the number of available workers.}
}
\value{
A data frame. The base data frame is the result combining the
\code{\link[bbr:nm_join]{bbr::nm_join()}} results for each chain submodel, collapsing across chains
with \code{point_fn}. \code{EPRED}, \code{IPRED}, \code{EWRES}, and \code{NPDE} values are replaced
with a simulated estimate, if requested by the corresponding argument.
}
\description{
\code{nm_join_bayes()} joins model output summaries to the input data. Underneath
this calls \code{\link[bbr:nm_join]{bbr::nm_join()}} on each chain submodel and then combines the
results, summarizing the table values across chains.

By default \code{nm_join_bayes()} replaces some table quantities with summaries of
\strong{simulated} values. It selects a subset of posterior samples and simulates
EPRED and IPRED with specified \pkg{mrgsolve} model. It also feeds the
simulated EPRED values to \pkg{npde} to calculate EWRES and NPDE values.
}
\details{
\subsection{Messages}{

\code{nm_join_bayes()} displays messages from \code{\link[bbr:nm_join]{bbr::nm_join()}} about the data and
table files being joined for the first chain. Messages for the subsequent
chains are omitted to avoid flooding the console with identical output.

As with \code{\link[bbr:nm_join]{bbr::nm_join()}}, you can suppress these messages by setting the
\code{bbr.verbose} option to \code{FALSE}.
}

\subsection{Parallel processing}{

To enable parallel processing for the EPRED and IPRED simulations, configure
\pkg{future} ahead of calling this function via \code{\link[future:plan]{future::plan()}}.
}

\subsection{Progress bar}{

The EPRED and IPRED simulations support displaying a progress bar via the
\pkg{progressr} package. The simplest way to enable a progress bar is to wrap
the \code{nm_join_bayes()} call inside a \code{\link[progressr:with_progress]{progressr::with_progress()}} call.
}
}
\seealso{
\code{\link[bbr:nm_join]{bbr::nm_join()}}, \link{bbr_nmbayes} for a high-level description of how
NONMEM Bayes models are structured in bbr
}
