% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit-model.R
\name{stan_submit_model}
\alias{stan_submit_model}
\alias{submit_model.bbi_stan_model}
\alias{submit_model.bbi_stan_gq_model}
\title{Submit model based on a \code{bbi_stan_model} object}
\usage{
\method{submit_model}{bbi_stan_model}(
  .mod,
  .bbi_args = NULL,
  .mode = c("local"),
  ...,
  .overwrite = NULL
)

\method{submit_model}{bbi_stan_gq_model}(
  .mod,
  .bbi_args = NULL,
  .mode = c("local"),
  ...,
  .overwrite = NULL
)
}
\arguments{
\item{.mod}{The model object to submit.}

\item{.bbi_args}{Unused argument (present for compatibility with
\code{\link[bbr:submit_model]{bbr::submit_model()}}).}

\item{.mode}{Mode of model submission. Stan models currently only support
local execution.}

\item{...}{Additional arguments (ignored for all Stan models).}

\item{.overwrite}{Logical to specify whether or not to overwrite existing
model output from a previous run. If \code{NULL}, the default, will defer to
setting in \code{.bbi_args} or \code{bbi.yaml}. If \emph{not} \code{NULL} will override any
settings in \code{.bbi_args} or \code{bbi.yaml}. \strong{The exception to this are
bootstrap runs (\code{bbi_nmboot_model} objects).} For bootstrap runs, this
defaults to \code{FALSE} and does \emph{not} respect any setting passed via
\code{.bbi_args} or a \code{bbi.yaml} config file. To overwrite existing bootstrap
output, a user must pass \code{TRUE} through this argument.}
}
\description{
Run the model via \pkg{cmdstanr}, selecting the \code{CmdStanModel} method based
the model type:
\itemize{
\item execute the model with \link[cmdstanr:model-method-sample]{$sample()} by
default
\item if object inherits from \code{bbi_stan_gq_model}, call
\link[cmdstanr:model-method-generate-quantities]{$generate_quantities()}
instead
}
}
\seealso{
\link{bbr_stan} for a high-level description of how Stan models are
structured
}
