% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanargs.R
\name{stanargs}
\alias{stanargs}
\alias{set_stanargs}
\alias{set_stanargs.bbi_stan_model}
\alias{set_stanargs.bbi_stan_gq_model}
\alias{get_stanargs}
\title{Set or get arguments to pass when submitting Stan model}
\usage{
set_stanargs(.mod, .stanargs, .clear = FALSE)

\method{set_stanargs}{bbi_stan_model}(.mod, .stanargs, .clear = FALSE)

\method{set_stanargs}{bbi_stan_gq_model}(.mod, .stanargs, .clear = FALSE)

get_stanargs(.mod)
}
\arguments{
\item{.mod}{A Stan model object.}

\item{.stanargs}{args passed through to \verb{$sample()} or
\verb{$generate_quantities()}.}

\item{.clear}{If \code{FALSE}, the default, add any new args and only overwrite
existing args that are passed in through \code{.stanargs}. If \code{TRUE}, overwrite
\emph{all} attached args so that only \code{.stanargs} will be attached.}
}
\description{
When \link[=stan_submit_model]{submit_model()} is called with a Stan model, it
collects the arguments from \verb{<run>-stanargs.R} and passes them to
\code{CmdStanModel}s \link[cmdstanr:model-method-sample]{$sample()} (for
\code{bbi_stan_model} objects) or
\link[cmdstanr:model-method-generate-quantities]{$generate_quantities()} (for
\code{bbi_stan_gq_model} objects).

\code{set_stanargs()} and \code{get_stanargs()} provide an interface for specifying and
inspecting the arguments attached to a model.
}
